Namespace `Chain`: more tests added.
